Revert "refactor: Change previous Modal to Window and update related components"
This reverts commit 9031831791b184146825f5eba628c608f4bdcbea.
